Move copied `prepare_program_extra_data`
